Navbar home bar working
