+++ constants/index.ts
@@ -1,12 +1,122 @@
 // NAVIGATION - Updated for eBay-style navbar
 export const NAV_LINKS = [
-  { href: '/sell', key: 'sell', label: 'Sell', type: 'button' },
-  { href: '/favorites', key: 'favorites', label: 'Favorites', type: 'icon', icon: 'heart' },
-  { href: '/inbox', key: 'inbox', label: 'Inbox', type: 'icon', icon: 'mail' },
-  { href: '/notifications', key: 'notifications', label: 'Notifications', type: 'icon', icon: 'bell' },
-  { href: '/profile', key: 'profile', label: 'Profile', type: 'icon', icon: 'user' },
-  { href: '/cart', key: 'cart', label: 'Cart', type: 'icon', icon: 'shopping-cart' },
-  { href: '/login', key: 'login', label: 'Login', type: 'icon', icon: 'profile' },
+  { href: '/sell', key: 'sell', label: 'Sell'},
+  { href: '/Favorites', key: 'favorites', label: 'Favorites'},
+  { href: '/inbox', key: 'inbox', label: 'Inbox'},
+  { href: '/notifications', key: 'notifications', label: 'Notifications' },
+  { href: '/profile', key: 'profile', label: 'Profile'},
+];
+
+// MARKETPLACE ITEMS
+export const SAMPLE_ITEMS = [
+  {
+    id: 1,
+    name: "Vintage Camera Canon AE-1",
+    price: 299.99,
+    image: "/api/placeholder/300/300",
+    location: "New York, NY",
+    shipping: "Free shipping",
+    condition: "Used - Excellent"
+  },
+  {
+    id: 2,
+    name: "MacBook Pro 13-inch M2 2022",
+    price: 1299.00,
+    image: "/api/placeholder/300/300",
+    location: "San Francisco, CA",
+    shipping: "$15.99 shipping",
+    condition: "Used - Very Good"
+  },
+  {
+    id: 3,
+    name: "Nike Air Jordan 1 Retro High",
+    price: 180.50,
+    image: "/api/placeholder/300/300",
+    location: "Chicago, IL",
+    shipping: "Free shipping",
+    condition: "New with box"
+  },
+  {
+    id: 4,
+    name: "Wooden Coffee Table Handmade",
+    price: 450.00,
+    image: "/api/placeholder/300/300",
+    location: "Austin, TX",
+    shipping: "Local pickup only",
+    condition: "New"
+  },
+  {
+    id: 5,
+    name: "iPhone 14 Pro Max 256GB",
+    price: 899.99,
+    image: "/api/placeholder/300/300",
+    location: "Los Angeles, CA",
+    shipping: "Free shipping",
+    condition: "Used - Good"
+  },
+  {
+    id: 6,
+    name: "Acoustic Guitar Yamaha FG830",
+    price: 199.99,
+    image: "/api/placeholder/300/300",
+    location: "Nashville, TN",
+    shipping: "$25.00 shipping",
+    condition: "Used - Very Good"
+  },
+  {
+    id: 7,
+    name: "Gaming Chair RGB LED",
+    price: 249.99,
+    image: "/api/placeholder/300/300",
+    location: "Seattle, WA",
+    shipping: "Free shipping",
+    condition: "New"
+  },
+  {
+    id: 8,
+    name: "Vintage Vinyl Record Collection",
+    price: 75.00,
+    image: "/api/placeholder/300/300",
+    location: "Detroit, MI",
+    shipping: "$12.99 shipping",
+    condition: "Used - Good"
+  },
+  {
+    id: 9,
+    name: "Mountain Bike Trek X-Caliber",
+    price: 650.00,
+    image: "/api/placeholder/300/300",
+    location: "Denver, CO",
+    shipping: "Local pickup only",
+    condition: "Used - Excellent"
+  },
+  {
+    id: 10,
+    name: "KitchenAid Stand Mixer",
+    price: 199.95,
+    image: "/api/placeholder/300/300",
+    location: "Boston, MA",
+    shipping: "Free shipping",
+    condition: "Used - Very Good"
+  },
+  {
+    id: 11,
+    name: "PlayStation 5 Console",
+    price: 499.99,
+    image: "/api/placeholder/300/300",
+    location: "Miami, FL",
+    shipping: "$19.99 shipping",
+    condition: "New in box"
+  },
+  {
+    id: 12,
+    name: "Designer Handbag Coach",
+    price: 125.00,
+    image: "/api/placeholder/300/300",
+    location: "Atlanta, GA",
+    shipping: "Free shipping",
+    condition: "Used - Good"
+  }
 ];
 
 // CAMP SECTION
