+++ src/app/Home/page.tsx
@@ -2,16 +2,12 @@ import { redirect } from "next/navigation";
 import MarketplaceGrid from "../../component/MarketplaceGrid";
 
 export default function Home() {
-  redirect("/Login");
-}
-
-
-export default function Home() {
   return (
-    <main>
+    <main className="p-8">
       <h1 className="font-bold text-blue-500 text-2xl">
-        <MarketplaceGrid />
+        Marketplace
       </h1>
+      <MarketplaceGrid />
     </main>
   );
 }
