+++ src/app/layout.tsx
@@ -1,6 +1,9 @@
 import type { Metadata } from "next";
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
+import Navbar from "../component/Navbar";
+import Sidebar from "../component/Sidebar";
+import Providers from "../providers/Providers";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -13,8 +16,8 @@ const geistMono = Geist_Mono({
 });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
-  description: "Generated by create next app",
+  title: "Coop",
+  description: "Marketplace for local communities",
 };
 
 export default function RootLayout({
@@ -27,7 +30,22 @@ export default function RootLayout({
       <body
         className={`${geistSans.variable} ${geistMono.variable} antialiased`}
       >
-        {children}
+        <Providers>
+          <Navbar />
+
+          {/* Main layout with sidebar and content */}
+          <div className="flex min-h-screen">
+            {/* Sidebar - positioned under navbar */}
+            <Sidebar />
+
+            {/* Main content area */}
+            <div className="flex-1 flex flex-col">
+              <main className="relative overflow-hidden">
+                {children}
+              </main>
+            </div>
+          </div>
+        </Providers>
       </body>
     </html>
   );
