+++ src/component/Button.tsx
@@ -8,9 +8,10 @@ type ButtonProps = {
 }
 
 const Button = ({ type, title, icon, variant }: ButtonProps) => {
+
   return (
     <button
-        className={`flexcenter gap-3 rounded-full border ${variant}`} 
+        className={`flexCenter gap-3 rounded-full border ${variant}`} 
         type={type}
     >
         {icon && <Image src={icon} alt={title} width={24} height={24} />}
