+++ src/component/Navbar.tsx
@@ -2,7 +2,7 @@
 
 import Image from "next/image"
 import Link from "next/link"
-import { NAV_LINKS } from "../constants"
+import { NAV_LINKS } from "../../constants"
 import { useState } from "react"
 
 // Icon components (you can replace these with your preferred icon library)
@@ -67,12 +67,12 @@ const Navbar = () => {
   return (
     <nav className="bg-white border-b border-gray-200 shadow-sm sticky top-0 z-50">
       {/* Main Navbar */}
-      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
+      <div className=" mx-auto px-4 sm:px-6 lg:px-8">
         <div className="flex items-center justify-between h-16">
           
           {/* Logo */}
           <div className="flex items-center">
-            <Link href="/" className="flex-shrink-0 flex items-center">
+            <Link href="/Home" className="flex-shrink-0 flex items-center">
               <Image 
                 src="/chicken.png" 
                 alt="logo" 
@@ -86,7 +86,7 @@ const Navbar = () => {
             </Link>
           </div>
 
-          {/* Search Bar - Desktop */}
+          {/* Search Bar */}
           <div className="flex-1 max-w-2xl mx-4 hidden md:block">
             <div className="relative">
               <input
@@ -129,65 +129,8 @@ const Navbar = () => {
               </Link>
             ))}
           </div>
-
-          {/* Mobile menu button */}
-          <div className="lg:hidden">
-            <button
-              onClick={() => setIsMobileMenuOpen(!isMobileMenuOpen)}
-              className="p-2 rounded-md text-gray-400 hover:text-gray-500 hover:bg-gray-100"
-            >
-              <MenuIcon />
-            </button>
-          </div>
-        </div>
-
-        {/* Search Bar - Mobile */}
-        <div className="md:hidden pb-4">
-          <div className="relative">
-            <input
-              type="text"
-              placeholder="Search for anything"
-              value={searchQuery}
-              onChange={(e) => setSearchQuery(e.target.value)}
-              className="w-full px-4 py-2 border border-gray-300 rounded-l-full focus:outline-none focus:ring-2 focus:ring-blue-500 focus:border-transparent"
-            />
-            <button className="absolute right-0 top-0 bottom-0 bg-blue-600 hover:bg-blue-700 text-white px-4 rounded-r-full">
-              <SearchIcon />
-            </button>
-          </div>
         </div>
       </div>
-
-      {/* Mobile Menu */}
-      {isMobileMenuOpen && (
-        <div className="lg:hidden bg-white border-t border-gray-200">
-          <div className="px-2 pt-2 pb-3 space-y-1 sm:px-3">
-            {NAV_LINKS.map((link) => (
-              <Link
-                key={link.key}
-                href={link.href}
-                className={`block px-3 py-2 rounded-md text-base font-medium transition-colors duration-200 ${
-                  link.type === 'button'
-                    ? 'bg-blue-600 text-white hover:bg-blue-700'
-                    : 'text-gray-700 hover:text-gray-900 hover:bg-gray-100'
-                } flex items-center space-x-2`}
-                onClick={() => setIsMobileMenuOpen(false)}
-              >
-                {link.icon && getIcon(link.icon)}
-                <span>{link.label}</span>
-                {link.key === 'cart' && (
-                  <span className="bg-red-500 text-white text-xs rounded-full h-5 w-5 flex items-center justify-center ml-auto">
-                    3
-                  </span>
-                )}
-                {link.key === 'notifications' && (
-                  <span className="bg-red-500 text-white text-xs rounded-full h-2 w-2 ml-auto"></span>
-                )}
-              </Link>
-            ))}
-          </div>
-        </div>
-      )}
     </nav>
   )
 }
